refactored the whole speed dial into its own component
